add ipc about io operation

+++ Motion_controller/src/process_comm/include/heartbeat_client.h
@@ -42,7 +42,7 @@ private:
 
     enum {CMD_HEARTBEAT = 0xA1,};
 
-    long computeTimeElapsed();
+    long long computeTimeElapsed();
 };
 
 }
