+++ Motion_controller/src/process_comm/include/interpreter_server.h
@@ -81,7 +81,7 @@ private:
     void copyRecvBufferToRequestData(void* request_data_ptr, int size);
     void copyResponseDataToSendBuffer(InterpreterServerCmd cmd_id, void* response_data_ptr, 
                                                 int response_data_size, int& send_buffer_size);
-    long computeTimeElapsed(struct timeval& current_time_val, struct timeval& last_time_val);
+    long long computeTimeElapsed(struct timeval& current_time_val, struct timeval& last_time_val);
     bool checkPublishCondition(long time_elapsed, int interval);
 
     // rpc request handler
